Clean: Relocate STEX CopyableByMemcpy
