+++ src/foundation/CopyableByMemcpy.h
